glycon/requirements.txt
# Core
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-Login==0.6.2
python-dotenv==1.0.0
pyOpenSSL==23.2.0

# Database
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.3

# Crypto
pycryptodome==3.18.0
cryptography==41.0.3

# Networking
requests==2.31.0
urllib3==2.0.4  # Required for Azure SDK
gunicorn==20.1.0  # Production WSGI server

# Azure Modules (Optional)
azure-identity==1.13.0
azure-mgmt-resource==23.0.1

# Utility
Werkzeug==2.3.7  # Pinned for Flask compatibility
Jinja2==3.1.2
click==8.1.4
python-dateutil==2.8.2
